+++ src/Captura.Core/Settings/Models/MouseClickSettings.cs
@@ -2,26 +2,8 @@
 
 namespace Captura
 {
-    public class MouseClickSettings : PropertyStore
+    public class MouseClickSettings : MouseOverlaySettings
     {
-        public bool Display
-        {
-            get => Get<bool>();
-            set => Set(value);
-        }
-
-        public int Radius
-        {
-            get => Get(25);
-            set => Set(value);
-        }
-
-        public Color Color
-        {
-            get => Get(Color.FromArgb(255, 235, 59));
-            set => Set(value);
-        }
-
         public Color RightClickColor
         {
             get => Get(Color.FromArgb(3, 169, 244));
@@ -31,18 +13,6 @@ namespace Captura
         public Color MiddleClickColor
         {
             get => Get(Color.FromArgb(76, 175, 80));
-            set => Set(value);
-        }
-
-        public int BorderThickness
-        {
-            get => Get<int>();
-            set => Set(value);
-        }
-
-        public Color BorderColor
-        {
-            get => Get(Color.Black);
             set => Set(value);
         }
     }